fix for pull request
